Move out tupleset ADT.
