+++ src/adt/Makefile
@@ -14,6 +14,7 @@ SRC += src/adt/edgeset.c
 SRC += src/adt/stateset.c
 SRC += src/adt/mappingset.c
 SRC += src/adt/transset.c
+SRC += src/adt/tupleset.c
 
 .for src in ${SRC:Msrc/adt/bitmap.c}
 CFLAGS.${src} += -I src # XXX: for internal.h
